requirements.txt
pyautogui==0.9.54
pygetwindow==0.0.9
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0